+++ src/hooks/renderJournalSheet/linkImages.js
@@ -1,6 +1,6 @@
 import utils from "../../utils.js";
 
-function linkImages(html) {
+export function linkImages(html) {
   if (!game.user.isGM) return;
   // does this functionality exist from anther module?
   const funcExists = utils.isModuleInstalledAndActive("vtta-dndbeyond") ||
@@ -13,34 +13,31 @@ function linkImages(html) {
   $(html)
     .find('div[data-edit="content"] img, div[data-edit="content"] video')
     .each((index, element) => {
-      const showPlayersButton = $("<a class='ddb-button'><i class='fas fa-eye'></i>&nbsp;Show Players</a>");
-      $(showPlayersButton).click(() => {
-        const src = $(element).attr("src");
-        game.socket.emit("module.ddb-importer", { sender: game.user.data._id, action: "showImage", src: src });
-      });
+      const showPlayersButton = $("<a class='ddbimporter-button'><i class='fas fa-eye'></i>&nbsp;Show Players</a>");
 
       $(element).wrap("<div class='ddbimporter-image-container'></div>");
       // show the button on mouseenter of the image
       $(element)
         .parent()
-        .mouseenter(function Hovering() {
+        .mouseenter(function addHover() {
+          // eslint-disable-next-line no-invalid-this
           $(this).append(showPlayersButton);
-          $(showPlayersButton).click(() => {
-            const src = $(element).attr("src");
+          $(showPlayersButton).click((event) => {
+            event.preventDefault();
+            event.stopPropagation();
             game.socket.emit("module.ddb-importer", {
               sender: game.user.data._id,
               action: "showImage",
-              src: src,
+              src: $(element).attr("src"),
               type: element.nodeName,
             });
           });
         });
       $(element)
         .parent()
-        .mouseleave(function Unhovering() {
+        .mouseleave(function removeHover() {
+          // eslint-disable-next-line no-invalid-this
           $(this).find("a").remove();
         });
     });
 }
-
-export default linkImages;