+++ src/index.js
@@ -1,5 +1,6 @@
 import {
   init,
+  onReady,
   onceReady,
   renderSidebarTab,
   renderJournalSheet,
@@ -10,6 +11,7 @@ CONFIG.debug.hooks = false;
 // register hooks
 Hooks.once("init", init);
 Hooks.once("ready", onceReady);
+Hooks.on("ready", onReady);
 Hooks.on("renderSidebarTab", renderSidebarTab);
 Hooks.on("renderJournalSheet", renderJournalSheet);
 Hooks.on("getSceneNavigationContext", extendSceneNavigationContext);