+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * The munch button is now at the top of the page (option to put at bottom in settings).
 * PC images can now be uploaded to separate folders to other images.
 * Image upload directory is now set and created to a sensible default.
+* Adventure Importer Journals with tables now offer a Roll on Table button.
 
 # 1.0.9